ttl system
